It's time to say goodbye to the CheckLyricTextTag.
